requirements.txt
# Core dependencies
neo4j>=5.15.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LLM integrations - OpenAI and Google Gemini (FREE APIs)
openai>=1.10.0
# New Google Generative AI package (recommended)
google-genai>=0.1.0
# Fallback to deprecated package if new one not available
google-generativeai>=0.3.0

# Data processing
python-dateutil>=2.8.2
tenacity>=8.2.3
jinja2>=3.1.2

# CLI
rich>=13.7.0
click>=8.1.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.3